agents/SummarizerAgent/requirements.txt
# SummarizerAgent Dependencies

# Google ADK and Vertex AI
google-adk>=1.18.0
vertexai>=1.38.0

# HTTP requests for MCP server communication
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

